#124: moved books to modules in store
