+++ src/store/index.js
@@ -2,12 +2,12 @@ import Vue from 'vue'
 import Vuex from 'vuex'
 import app from './modules/app'
 import portfolio from './modules/portfolio'
-import books from './books'
 import createPersistedState from 'vuex-persistedstate'
 import vuexBitsharesModules from 'vuex-bitshares'
 import Cookies from 'js-cookie'
 import paths from './cachedPaths'
 import markets from './modules/markets'
+import books from './modules/books'
 
 Vue.use(Vuex)
